feat(pwa): unified sign-in / sign-up page with redirect-back
Restricted pages used to bounce anonymous visitors to a hard-coded
`/signin`, dropping whatever URL they were originally trying to view.
And signing in always landed on `/account`, even when the user had
deep-linked to `/projects/abc` first. Two changes:

1. Restricted pages now redirect to `/signin?next=<original-path>`
   via a shared `signinHrefForCurrent()` helper. The auth screens
   read `next`, validate it (must be root-relative, no `//host` or
   absolute URLs — guarded by `safeNextPath()`), and route the user
   back there after login. Same `next` is preserved through the
   sign-up → registered → sign-in handshake, and an open-redirect
   attempt falls back to `/account`.

2. `/signin` and `/signup` now share a single tabbed `AuthCard`
   component. Both URLs work and deep-linking still does the right
   thing: `/signin` defaults to the Sign In tab, `/signup` to the
   Sign Up tab, and switching tabs is instant — no navigation. The
   existing `?invite=<token>` flow keeps working because the invite
   token is read inside `SignUpForm` regardless of which URL the
   user landed on.

If a logged-in user lands on /signin or /signup (old bookmark), the
card silently bounces them to `next` (or `/account`).

Adds `pwa/components/ui/tabs.tsx` (shadcn primitive), pulls the
existing forms into reusable `SignInForm` / `SignUpForm` components,
and ships an `auth-redirect.spec.js` E2E covering the redirect-back
path, the sign-up→sign-in flow with a preserved `next`, and the
open-redirect rejection.
