+++ pwa/pages/tags.tsx
@@ -3,6 +3,7 @@ import { useRouter } from "next/router";
 import { FormEvent, useCallback, useEffect, useState } from "react";
 import { useAuth } from "@/contexts/AuthContext";
 import { ENTRYPOINT } from "@/config/entrypoint";
+import { signinHrefForCurrent } from "@/lib/authRedirect";
 import MarkdownEditor from "@/components/editor/MarkdownEditor";
 import MarkdownView from "@/components/editor/MarkdownView";
 import { Alert, AlertDescription } from "@/components/ui/alert";
@@ -48,7 +49,7 @@ const Tags = () => {
 
   useEffect(() => {
     if (!authLoading && !isAuthenticated) {
-      router.push("/signin");
+      router.replace(signinHrefForCurrent(router.asPath));
     }
   }, [authLoading, isAuthenticated, router]);
